rustc: move Instance::fn_sig to ty::layout and privatize it.
